Re-organise RDD
